requirements.txt
asgiref==3.11.0
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
numpy==2.4.0
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.11
pyshp==3.0.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
